# 🧠⚡ Biomimetic Neuromorphic Foundation - 1000x Amplification Through Cognitive Architecture