… docker-compose.biomimetic.yml
… docs/BIOMIMETIC_ARCHITECTURE.md
… src/biomimetic/memory_router.py